Price Crawler and Product Caching
Implemented the price crawler which will get all products stored in the
database and get their current price on Amazon. This still needs some
more refinement as some product prices still aren’t being returned
(mainly for ones that are only being sold as refurbished).

When an AmazonItem is created it first tries to grab the product form
the database, if it’s not found then we get it from the Amazon API and
store the product in the database for future use.